Further improved things with an enumeration based PageTable descent.  It's better for memory overhead and hardly noticeable for any slowdown.
I made some phases of the process analysis parallelized again so this more than made up for the enumeration code rather than a straight loop.

General cleanups for messy stuff aswell.

Testing on 64GB memory dumps is now good, no issues.  If any pop up down the road (testing on 512GB box is planned in a couple weeks) now that the page table parsing is an enumeration we can chunk/stream the input to avoid any issues with large files.